+++ Plugins/SimpleQuest/Source/SimpleQuestEditor/Private/Nodes/Slate/SGraphNode_QuestlineQuest.cpp
@@ -121,9 +121,7 @@ void SGraphNode_QuestlineQuest::UpdateGraphNode()
 	// ── Inner content ──────────────────────────────────────────
 	FGraphNodeMetaData TagMeta(TEXT("Graphnode"));
 	PopulateMetaTag(&TagMeta);
-
-	this->ContentScale.Bind(this, &SGraphNode::GetContentScale);
-
+	
 	TSharedPtr<SVerticalBox> InnerVerticalBox = SNew(SVerticalBox)
 		+ SVerticalBox::Slot().AutoHeight().HAlign(HAlign_Fill).VAlign(VAlign_Top)
 		.Padding(Settings->GetNonPinNodeBodyPadding())